+++ .idea/dictionaries/huqingyang.xml
@@ -3,6 +3,7 @@
     <words>
       <w>classnames</w>
       <w>formily</w>
+      <w>hotarea</w>
       <w>runtimes</w>
       <w>vize</w>
       <w>vizerc</w>
