+++ packages/editor/src/states/components.ts
@@ -10,6 +10,7 @@ import {
   componentEventDepsMap,
   ComponentIndex,
   createComponentInstance,
+  deleteComponentNode,
   deleteCurrentPageComponentIndex,
   deletePageComponentInstanceIndexMap,
   DepsType,
@@ -112,6 +113,7 @@ export class ComponentsStore {
     selectStore.selectPage(selectStore.pageIndex);
     eventStore.deleteDepsEventInstances(DepsType.Component, key);
     componentEventDepsMap.deleteEventDepsMap(key);
+    deleteComponentNode(key);
   };
 
   @action
