+++ packages/editor/src/types/component.ts
@@ -2,9 +2,9 @@ import * as React from 'react';
 import { MaterialsForm, MaterialsInfo } from './materials';
 import { MaterialsCustomEvent } from './events';
 import { EventInstance } from './events';
-import { CommonStyleMeta, Percent } from './styles';
+import { CommonStyleMeta } from './styles';
 import { Maybe } from './helper';
-import { GlobalMeta } from './global';
+import { HotArea } from './hotArea';
 
 export interface MaterialsComponentMeta {
   identityName: string;
@@ -35,30 +35,6 @@ export interface ComponentSize {
   height: number;
 }
 
-export interface HotAreaPosition {
-  x: Percent;
-  y: Percent;
-}
-
-export interface HotAreaSize {
-  width: Percent;
-  height: Percent;
-}
-
-export interface HotAreaParamsData {
-  key: number;
-  position: HotAreaPosition;
-  size: HotAreaSize;
-  parent: MaterialsComponentMeta;
-}
-
-export interface HotArea {
-  key: number;
-  position: HotAreaPosition;
-  size: HotAreaSize;
-  events: EventInstance[];
-}
-
 export interface ComponentInstance {
   key: Readonly<number>;
   component: Readonly<string>;
@@ -80,43 +56,6 @@ export interface ComponentProps extends Pick<ComponentInstance, 'data' | 'style'
   componentKey: Readonly<number>;
   instance: ComponentInstance;
   hotAreas: Maybe<React.ReactElement>;
-}
-
-interface EditorMaterialCallbackParams {
-  global: object;
-  meta: GlobalMeta;
-  originalEvent?: React.SyntheticEvent;
-}
-export type ComponentsMapType = Map<number, HTMLDivElement>;
-export interface EditorMaterialPluginParams extends EditorMaterialCallbackParams {
-  key: number;
-  data: object;
-  on: Function;
-  componentsMap: ComponentsMapType;
-}
-
-export interface EditorMaterialActionParams extends EditorMaterialCallbackParams {
-  key: number;
-  data?: object;
-  triggerEventType: string;
-  trigger: {
-    type: 'component' | 'hotarea';
-    component: MaterialsComponentMeta;
-    hotarea?: HotAreaParamsData;
-  };
-  target?: MaterialsComponentMeta;
-  componentsMap: ComponentsMapType;
-}
-
-export interface EditorMaterialPluginEventListenerCallbackParams extends EditorMaterialCallbackParams {
-  triggerEventType: string;
-  trigger: {
-    type: 'component' | 'hotarea';
-    component: MaterialsComponentMeta;
-    hotarea?: HotAreaParamsData;
-  };
-  component: MaterialsComponentMeta;
-  hotArea?: HotAreaParamsData;
 }
 
 export type MaterialsComponent = React.ComponentType<ComponentProps>;
