+++ packages/editor/src/utils/index.ts
@@ -1,7 +1,8 @@
+export * from './indexMap';
 export * from './loader';
+export * from './nodeMap';
+export * from './color';
 export * from './common';
-export * from './componentNodes';
-export * from './hotAreaNodes';
 export * from './contextMenu';
 export * from './create';
 export * from './depsMap';
@@ -9,9 +10,7 @@ export * from './dom';
 export * from './dsl';
 export * from './eventEmiter';
 export * from './eventHelper';
-export * from './eventParams';
 export * from './eventProxy';
-export * from './indexMap';
 export * from './init';
 export * from './inject';
 export * from './is';
@@ -19,5 +18,3 @@ export * from './key';
 export * from './materialsMap';
 export * from './request';
 export * from './style';
-export * from './color';
-export * from './eventListener';
