+++ packages/editor/src/utils/nodeMap/hotAreaNodes.ts
@@ -1,10 +1,10 @@
 import { Maybe } from 'types';
-import { getHotAreaId } from './common';
+import { getHotAreaId } from '../common';
 
 export const hotAreaNodeMap = new Map<number, HTMLDivElement>();
 
 export function setHotAreaNode(key: number, node: HTMLDivElement) {
-  hotAreaNodeMap.set(key, node);
+  return hotAreaNodeMap.set(key, node);
 }
 
 export function deleteHotAreaNode(key: number) {
@@ -14,9 +14,3 @@ export function deleteHotAreaNode(key: number) {
 export function getHotAreaNode(key: number): Maybe<HTMLDivElement> {
   return hotAreaNodeMap.get(key) || (document.getElementById(getHotAreaId(key)) as Maybe<HTMLDivElement>);
 }
-
-Object.defineProperty(window, '__hotAreaInstanceNodeMap', {
-  get() {
-    return hotAreaNodeMap;
-  },
-});
