+++ packages/editor/src/utils/nodeMap/index.ts
@@ -1,0 +1,2 @@
+export * from './componentNodes';
+export * from './hotAreaNodes';
